chore(refactor): Extract and refactor middlewares
